requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.6
dataclasses==0.6
dateparser==1.0.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
distro==1.5.0
Django==3.1.7
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-categories==1.8
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-crispy-forms==1.11.1
django-formtools==2.2
django-froala-editor==3.2.2
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-robots==4.0
django-sekizai==2.0.0
django-six==1.0.4
django-summernote==0.8.11.6
django-taggit==1.3.0
django-tinymce==3.2.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
entrypoints==0.3
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
ipaddr==2.2.0
ipykernel==5.5.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.5.2
lockfile==0.12.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mediapipe==0.8.3.1
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.1
pep517==0.9.1
pickleshare==0.7.5
Pillow==8.1.0
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.7
psycopg2==2.8.6
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycountry==20.7.3
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.27
python-binance==1.0.10
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
unicode-slugify==0.1.3
Unidecode==1.2.0
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
